automation-scripts/openclipart-scraper/requirements.txt
aiohttp>=3.9
beautifulsoup4>=4.11
pyyaml>=6.0
